+++ cpp/apps/Open3DViewer/Open3DViewer.cpp
@@ -29,8 +29,8 @@
 #include <string>
 
 #include "open3d/Open3D.h"
-#include "open3d/visualization/Visualizer/GuiVisualizer.h"
 #include "open3d/visualization/gui/Native.h"
+#include "open3d/visualization/visualizer/GuiVisualizer.h"
 
 using namespace open3d;
 using namespace open3d::geometry;
