+++ cpp/open3d/Open3D.h.in
@@ -69,14 +69,6 @@
 #include "open3d/visualization/Utility/DrawGeometry.h"
 #include "open3d/visualization/Utility/SelectionPolygon.h"
 #include "open3d/visualization/Utility/SelectionPolygonVolume.h"
-#include "open3d/visualization/Visualizer/ViewControl.h"
-#include "open3d/visualization/Visualizer/ViewControlWithCustomAnimation.h"
-#include "open3d/visualization/Visualizer/ViewControlWithEditing.h"
-#include "open3d/visualization/Visualizer/Visualizer.h"
-#include "open3d/visualization/Visualizer/VisualizerWithCustomAnimation.h"
-#include "open3d/visualization/Visualizer/VisualizerWithEditing.h"
-#include "open3d/visualization/Visualizer/VisualizerWithKeyCallback.h"
-#include "open3d/visualization/Visualizer/VisualizerWithVertexSelection.h"
 #include "open3d/visualization/gui/Application.h"
 #include "open3d/visualization/gui/Button.h"
 #include "open3d/visualization/gui/Checkbox.h"
@@ -95,6 +87,14 @@
 #include "open3d/visualization/gui/TextEdit.h"
 #include "open3d/visualization/gui/Theme.h"
 #include "open3d/visualization/gui/Window.h"
+#include "open3d/visualization/visualizer/ViewControl.h"
+#include "open3d/visualization/visualizer/ViewControlWithCustomAnimation.h"
+#include "open3d/visualization/visualizer/ViewControlWithEditing.h"
+#include "open3d/visualization/visualizer/Visualizer.h"
+#include "open3d/visualization/visualizer/VisualizerWithCustomAnimation.h"
+#include "open3d/visualization/visualizer/VisualizerWithEditing.h"
+#include "open3d/visualization/visualizer/VisualizerWithKeyCallback.h"
+#include "open3d/visualization/visualizer/VisualizerWithVertexSelection.h"
 
 // clang-format off
 @BUILD_AZURE_KINECT_COMMENT@#include "open3d/io/sensor/azure_kinect/AzureKinectRecorder.h"
