+++ cpp/open3d/visualization/Shader/Simple2DShader.cpp
@@ -28,7 +28,7 @@
 
 #include "open3d/visualization/Shader/Shader.h"
 #include "open3d/visualization/Utility/SelectionPolygon.h"
-#include "open3d/visualization/Visualizer/RenderOptionWithEditing.h"
+#include "open3d/visualization/visualizer/RenderOptionWithEditing.h"
 
 namespace open3d {
 namespace visualization {
