+++ cpp/open3d/visualization/Utility/DrawGeometry.cpp
@@ -26,15 +26,15 @@
 
 #include "open3d/visualization/Utility/DrawGeometry.h"
 
-#include "open3d/visualization/Visualizer/GuiVisualizer.h"
-#include "open3d/visualization/Visualizer/ViewControlWithCustomAnimation.h"
-#include "open3d/visualization/Visualizer/ViewControlWithEditing.h"
-#include "open3d/visualization/Visualizer/Visualizer.h"
-#include "open3d/visualization/Visualizer/VisualizerWithCustomAnimation.h"
-#include "open3d/visualization/Visualizer/VisualizerWithEditing.h"
-#include "open3d/visualization/Visualizer/VisualizerWithKeyCallback.h"
-#include "open3d/visualization/Visualizer/VisualizerWithVertexSelection.h"
 #include "open3d/visualization/gui/Application.h"
+#include "open3d/visualization/visualizer/GuiVisualizer.h"
+#include "open3d/visualization/visualizer/ViewControlWithCustomAnimation.h"
+#include "open3d/visualization/visualizer/ViewControlWithEditing.h"
+#include "open3d/visualization/visualizer/Visualizer.h"
+#include "open3d/visualization/visualizer/VisualizerWithCustomAnimation.h"
+#include "open3d/visualization/visualizer/VisualizerWithEditing.h"
+#include "open3d/visualization/visualizer/VisualizerWithKeyCallback.h"
+#include "open3d/visualization/visualizer/VisualizerWithVertexSelection.h"
 
 namespace open3d {
 namespace visualization {
