+++ cpp/open3d/visualization/visualizer/Visualizer.cpp
@@ -24,7 +24,7 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "open3d/visualization/Visualizer/Visualizer.h"
+#include "open3d/visualization/visualizer/Visualizer.h"
 
 #include "open3d/geometry/TriangleMesh.h"
