+++ cpp/open3d/visualization/visualizer/VisualizerWithCustomAnimation.h
@@ -26,7 +26,7 @@
 
 #pragma once
 
-#include "open3d/visualization/Visualizer/Visualizer.h"
+#include "open3d/visualization/visualizer/Visualizer.h"
 
 namespace open3d {
 namespace visualization {
