+++ cpp/pybind/visualization/renderoption.cpp
@@ -24,7 +24,7 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "open3d/visualization/Visualizer/RenderOption.h"
+#include "open3d/visualization/visualizer/RenderOption.h"
 #include "open3d/io/IJsonConvertibleIO.h"
 
 #include "pybind/docstring.h"
