+++ cpp/pybind/visualization/visualizer.cpp
@@ -24,11 +24,11 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "open3d/visualization/Visualizer/Visualizer.h"
+#include "open3d/visualization/visualizer/Visualizer.h"
 #include "open3d/geometry/Image.h"
-#include "open3d/visualization/Visualizer/VisualizerWithEditing.h"
-#include "open3d/visualization/Visualizer/VisualizerWithKeyCallback.h"
-#include "open3d/visualization/Visualizer/VisualizerWithVertexSelection.h"
+#include "open3d/visualization/visualizer/VisualizerWithEditing.h"
+#include "open3d/visualization/visualizer/VisualizerWithKeyCallback.h"
+#include "open3d/visualization/visualizer/VisualizerWithVertexSelection.h"
 
 #include "pybind/docstring.h"
 #include "pybind/visualization/visualization.h"
